Move to widgets.py file
